requirements.txt
flask==2.3.3
google-genai>=1.0.0
pyttsx3==2.90
SpeechRecognition==3.10.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pygame==2.5.2
pydub==0.25.1
pyaudio==0.2.13
numpy==1.26.3
Pillow>=10.0.0
websockets>=11.0.0
asyncio>=3.4.3
vosk==0.3.45
TTS==0.22.0
torch>=2.0.0
